Extract conversion function
